Re-enabled additional test
